+++ lib/sequentia/datasets/digits.py
@@ -1,5 +1,5 @@
 from pkg_resources import resource_filename
-from typing import Sequence
+from typing import Iterable
 from operator import itemgetter
 
 import numpy as np
@@ -10,19 +10,19 @@ from sequentia.utils.validation import _Validator
 from sequentia.utils.decorators import _validate_params
 
 class _DigitsValidator(_Validator):
-    numbers: Sequence[conint(ge=0, le=9)] = list(range(10))
+    digits: Iterable[conint(ge=0, le=9)] = list(range(10))
 
-    @validator('numbers')
-    def check_numbers(cls, value):
+    @validator('digits')
+    def check_digits(cls, value):
         value = list(value)
         if len(set(value)) < len(value):
-            raise ValueError('Expected numbers to be unique')
+            raise ValueError('Expected digits to be unique')
         return value
 
 @_validate_params(using=_DigitsValidator)
 def load_digits(
     *,
-    numbers: Sequence[int] = list(range(10)), 
+    digits: Iterable[int] = list(range(10)),
 ) -> SequentialDataset:
     """Loads MFCC features of spoken digit audio samples from the Free Spoken Digit Dataset.
 
@@ -31,7 +31,7 @@ def load_digits(
 
     This version consists of 13 MFCC features of 50 recordings for each digit by 6 individual speakers.
 
-    :param numbers: Subset of digits to include in the dataset.
+    :param digits: Subset of digits to include in the dataset.
     :return: A dataset object representing the loaded digits.
     """
     # Load the dataset from compressed numpy file
@@ -41,7 +41,7 @@ def load_digits(
     X, y, lengths = itemgetter('X', 'y', 'lengths')(data)
 
     # Select and create a Dataset only with sequences having the specified labels
-    idx = np.argwhere(np.isin(y, numbers)).flatten()
+    idx = np.argwhere(np.isin(y, digits)).flatten()
     ranges = SequentialDataset._get_idxs(lengths)[idx]
     return SequentialDataset(
         np.vstack([x for x in SequentialDataset._iter_X(X, ranges)]),